+++ packages/Router/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+### Addded
+- window event for not found routes
+
+### Breaking Changes
+- renamed element class names to kebab case (mainly effects DomxRoute -> domx-route)
+
+
 ## [0.1.2] 2021-09-22
 ### Updated
 - dependencies